+++ src/ConfigParsing.cs
@@ -13,7 +13,7 @@ using ThreeDISevenZeroR.UnityGifDecoder;
 using ThreeDISevenZeroR.UnityGifDecoder.Model;
 using UnityEngine;
 using UnityEngine.Playables;
-using static RumbleModdingAPI.Calls;
+using RumbleModdingAPI.RMAPI;
 
 namespace RumblePhotoAlbum;
 
@@ -388,7 +388,7 @@ public partial class MainClass : MelonMod
 
         int pictureLayer = pictureData.visible?
             LayerMask.NameToLayer("UI") // No collision, visible
-            : LayerMask.NameToLayer("PlayerFade"); // No collision, invisible
+            : LayerMask.NameToLayer("ScreenFade"); // No collision, invisible
 
         float aspectRatio = (float)imageTexture.height / imageTexture.width;
         if (pictureData.width == 0 && pictureData.height == 0)
@@ -429,6 +429,8 @@ public partial class MainClass : MelonMod
         Renderer frameRenderer = frame.GetComponent<Renderer>();
         frameRenderer.material.shader = Shader.Find("Shader Graphs/RUMBLE_Prop");
         frameRenderer.material.SetColor("_Overlay", pictureData.color);
+        frameRenderer.material.SetFloat("_Metalic", pictureData.metallic);
+
 
         // Create quad with the image on it
         GameObject quad = GameObject.CreatePrimitive(PrimitiveType.Quad);
@@ -446,6 +448,7 @@ public partial class MainClass : MelonMod
         Renderer quadRenderer = quad.GetComponent<Renderer>();
         quadRenderer.material.shader = Shader.Find("Shader Graphs/RUMBLE_Prop");
         quadRenderer.material.SetTexture("_Albedo", imageTexture);
+        quadRenderer.material.SetFloat("_Metalic", 0);
 
         // Make the picture interactable
         pictureData.obj = obj;
@@ -499,7 +502,7 @@ public partial class MainClass : MelonMod
         var renderer = CreatePictureBlock(ref pictureData, parent, firstFrame.texture);
 
         // add indicator that the gif is loading
-        GameObject loadingText = Calls.Create.NewText();
+        GameObject loadingText = Create.NewText();
         TextMeshPro component = loadingText.GetComponent<TextMeshPro>();
         component.text = "Loading...";
         component.fontSize = 1f;
@@ -740,7 +743,7 @@ public partial class MainClass : MelonMod
         // Load image into Texture2D with alpha channel
         byte[] data = File.ReadAllBytes(path);
         Texture2D input = new Texture2D(2, 2, TextureFormat.RGBA32, false);
-        input.LoadImage(data);
+        input.LoadImage( data);
 
         // Get all pixels of the image at once
         Color[] inputPixels = input.GetPixels();