+++ src/main.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 using System.IO;
 using MelonLoader;
 using Newtonsoft.Json.Linq;
-using RumbleModdingAPI;
+using RumbleModdingAPI.RMAPI;
 using UnityEngine;
 using Il2CppRUMBLE.Interactions.InteractionBase;
 
@@ -12,7 +12,7 @@ namespace RumblePhotoAlbum;
 public static class BuildInfo
 {
     public const string ModName = "RumblePhotoAlbum";
-    public const string ModVersion = "1.2.3";
+    public const string ModVersion = "1.3.1";
     public const string Description = "Decorate your environment with framed pictures";
     public const string Author = "Kalamart";
     public const string Company = "";
@@ -33,6 +33,7 @@ public partial class MainClass : MelonMod
         public float height = 0;
         public float padding = defaultPadding;
         public float thickness = defaultThickness;
+        public float metallic = defaultMetallicness;
         public Color color = defaultColor;
         public bool alpha = false;
         public bool visible = true;
@@ -45,6 +46,7 @@ public partial class MainClass : MelonMod
     // variables
     protected static float defaultSize = 0.5f; // Default size of the frame (width or height depending on the orientation)
     protected static float defaultThickness = 0.01f; // Default thickness of the frame
+    protected static float defaultMetallicness = 0; // Default metallicness of the frame
     protected static float defaultPadding = 0.01f; // Default frame padding around the picture
     protected static Color defaultColor = new Color(0.48f, 0.80f, 0.76f); // Rumble gym green as default frame color
 
@@ -57,7 +59,7 @@ public partial class MainClass : MelonMod
     protected static bool buttonsVisibility = true; // Whether the buttons are visible on top of the held picture
     
     protected static GameObject photoAlbum = null; // Parent object for all framed pictures
-    protected static string currentScene = "";
+    protected static string currentScene = "Loader";
     private static bool flatlandFound = false;
 
     private static List<PictureData> PicturesList = null;
@@ -98,8 +100,8 @@ public partial class MainClass : MelonMod
     public override void OnLateInitializeMelon()
     {
         EnsureUserDataFolders();
-        Calls.onMapInitialized += OnMapInitialized;
-        Calls.onMyModsGathered += CheckMods;
+        Actions.onMapInitialized += OnMapInitialized;
+        Actions.onMyModsGathered += CheckMods;
     }
 
     /**
@@ -121,10 +123,16 @@ public partial class MainClass : MelonMod
     * Called when the full map is initialized, and RMAPI calls can be used safely.
     * </summary>
     */
-    private void OnMapInitialized()
+    private void OnMapInitialized(string sceneName)
     {
+        currentScene = sceneName;
+        if (sceneName == "Gym" && flatlandFound)
+        {
+            MelonCoroutines.Start(ListenForFlatLandButton());
+        }
+
         initializeInteractionObjects();
-        MelonCoroutines.Start(LoadAlbum(currentScene));
+        MelonCoroutines.Start(LoadAlbum(sceneName));
     }
 
     /**
@@ -134,15 +142,9 @@ public partial class MainClass : MelonMod
     */
     public override void OnSceneWasLoaded(int buildIndex, string sceneName)
     {
-        currentScene = sceneName;
         if (sceneName == "Loader")
         {
             InitModUI();
-            return;
-        }
-        else if (sceneName == "Gym"  && flatlandFound)
-        {
-            MelonCoroutines.Start(ListenForFlatLandButton());
         }
     }