+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.3.1
+- Fixed metallic aspect
+- Added option to configure the metallic aspect of the frame
+- Added requirement for MelonLoader 0.7.2
+
+# Version 1.3.0
+- Fixed the mod for v0.5.0.1
+
+# Version 1.2.4
+- Fixed action button visibility option not working
+
 # Version 1.2.3
 - Fixed bug with GIF not playing sometimes